+++ src/data/shortlist.h
@@ -260,6 +260,14 @@ public:
       for(auto& it : data_[i])
         indexSet.insert(it.first);
     }
+    // Ensure that the generated vocabulary items from a shortlist are a multiple-of-eight
+    // This is necessary until intgemm supports non-multiple-of-eight matrices.
+    // TODO better solution here? This could potentially be slow.
+    WordIndex i = static_cast<WordIndex>(firstNum_);
+    while (indexSet.size() % 8 != 0) {
+      indexSet.insert(i);
+      i++;
+    }
 
     // turn into vector and sort (selected indices)
     std::vector<WordIndex> indices(indexSet.begin(), indexSet.end());