+++ src/tensors/backend.h
@@ -10,10 +10,7 @@ protected:
   DeviceId deviceId_;
   size_t seed_;
   Ptr<RandomGenerator> randomGenerator_;
-
-  // global clipping value for matrix-multiplies, should soon be removed.
-  float clipValue_{0.f};
-
+  
 public:
   Backend(DeviceId deviceId, size_t seed)
       : deviceId_(deviceId), seed_(seed), randomGenerator_(createRandomGenerator(seed, deviceId)) {}
@@ -24,14 +21,6 @@ public:
   // for GPU only, calls cudaSetDevice, does nothing on CPU. Maybe change name.
   virtual void setDevice() = 0;
   virtual void synchronize() = 0;
-
-  virtual void setClip(float clipValue) { clipValue_ = clipValue; }
-  float getClip() { return clipValue_; }
-
-  // for CPU, sets to use optimized code for inference.
-  // for GPU, this is invalid. for gpu, isOptimized() function always returns false.
-  virtual void setOptimized(bool optimize) = 0;
-  virtual bool isOptimized() = 0;
 };
 
 Ptr<Backend> BackendByDeviceId(DeviceId deviceId, size_t seed);